+++ src/SimpleSocket.h
@@ -454,7 +454,7 @@ public:
 
     /// Returns the port number on which the client is connected.
     ///  @return client port number.
-    int16 GetClientPort() {
+    uint16 GetClientPort() {
         return m_stClientSockaddr.sin_port;
     };
 
@@ -466,35 +466,35 @@ public:
 
     /// Returns the port number on which the server is connected.
     ///  @return server port number.
-    int16 GetServerPort() {
+    uint16 GetServerPort() {
         return ntohs(m_stServerSockaddr.sin_port);
     };
 
     /// Get the TCP receive buffer window size for the current socket object.
     /// <br><br>\b NOTE: Linux will set the receive buffer to twice the value passed.
     ///  @return zero on failure else the number of bytes of the TCP receive buffer window size if successful.
-    uint16 GetReceiveWindowSize() {
+    uint32 GetReceiveWindowSize() {
         return GetWindowSize(SO_RCVBUF);
     };
 
     /// Get the TCP send buffer window size for the current socket object.
     /// <br><br>\b NOTE: Linux will set the send buffer to twice the value passed.
     ///  @return zero on failure else the number of bytes of the TCP receive buffer window size if successful.
-    uint16 GetSendWindowSize() {
+    uint32 GetSendWindowSize() {
         return GetWindowSize(SO_SNDBUF);
     };
 
     /// Set the TCP receive buffer window size for the current socket object.
     /// <br><br>\b NOTE: Linux will set the receive buffer to twice the value passed.
     ///  @return zero on failure else the number of bytes of the TCP send buffer window size if successful.
-    uint16 SetReceiveWindowSize(uint16 nWindowSize) {
+    uint32 SetReceiveWindowSize(uint32 nWindowSize) {
         return SetWindowSize(SO_RCVBUF, nWindowSize);
     };
 
     /// Set the TCP send buffer window size for the current socket object.
     /// <br><br>\b NOTE: Linux will set the send buffer to twice the value passed.
     ///  @return zero on failure else the number of bytes of the TCP send buffer window size if successful.
-    uint16 SetSendWindowSize(uint16 nWindowSize) {
+    uint32 SetSendWindowSize(uint32 nWindowSize) {
         return SetWindowSize(SO_SNDBUF, nWindowSize);
     };
 
@@ -523,11 +523,11 @@ protected:
 private:
     /// Generic function used to get the send/receive window size
     ///  @return zero on failure else the number of bytes of the TCP window size if successful.
-    uint16 GetWindowSize(uint32 nOptionName);
+    uint32 GetWindowSize(uint32 nOptionName);
 
     /// Generic function used to set the send/receive window size
     ///  @return zero on failure else the number of bytes of the TCP window size if successful.
-    uint16 SetWindowSize(uint32 nOptionName, uint32 nWindowSize);
+    uint32 SetWindowSize(uint32 nOptionName, uint32 nWindowSize);
 
 
     /// Attempts to send at most nNumItem blocks described by sendVector
